Add missing catalog entries.
Add catalog entry for namespace "http://www.jboss.com/xml/ns/javaee" so
that others can refer to it without specifying schema location.

Add catalog entry for namespace "urn:clustering:1.0" which is mentioned
for example in
https://docs.jboss.org/author/display/AS71/jboss-ejb3.xml+Reference but
missing in this plugin and in the JBoss EAP 6 distribution as far as I
see.