JBIDE-21949 - Allow user to turn on/off showing console when output changes
Changing the code to show the message console when it is created only,
then offering two methods to write lines in the shell and activate the
view or not (default is 'true').

Also, upgrading the project's execution environment to Java 1.7 to take
advantage of the 'try-with-resources' statement to make sure that the
outputstream is closed upon writting.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>