JBIDE-22227 - Ensure hotcode replace functions on local wildfly.
This will adjust the behavior of the 'restart module' option in the
hotcode replace dialog to not only republish the modules, but then
wait for the module publish to complete, remove breakpoints,
execute a garbage collection request, and then re-instate the
breakpoints.

Special thanks to Thomas Maeder for the solution.