Jbide 26598 fix jmx navigator local j vm with jdk9+everywhere improved withjbosstools common to modify too (#620)
* JBIDE-26598 - fix JMX navigator local processes with JDK9+ Eclipse
instance targeting a JDK9+ process

with JDK9+, there is no management-agent.jar put aside in lib folder of
the JDk home. It is available by default.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>

* JBIDE-26598 - check for Jigsaw on the targeted VM (and not the eclipse
instance)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>

* For Java9+ processes, don't load agent but rather start local management agent

Signed-off-by: Jeff MAURY <jmaury@redhat.com>

* JBIDE-26598 - workaround to avoid exception in UI

bootclasspath is not supported with Java 11. When selecting a connected
process running on Java 11 in JMX Navigator, the if properties view is
opened, there was UI popup error. Now it is only logging as Warning
waiting for a proper fix. The property in UI will appear blank, without
value.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>